fix: fix integration test for creating user

+++ src/test/api/create-user.test.ts
@@ -6,8 +6,6 @@ import { UserRecord } from "../../db/schema";
 import { deleteUser } from "../../db/queries/users";
 
 describe("/api/users", () => {
-    
-
     it("should create user record in db and return UserRecord", async () => {
         const payload: UserRequestData = { userName: "guest123", password: "passwordpassword" };
 
@@ -28,7 +26,39 @@ describe("/api/users", () => {
         }
 
         expect(response.status).toBe(201)
-        expect(response.body).toBe("UserRecord");
+        expect(user.userName).toBe(payload.userName);
+        for (const property in user) {
+            expect(property).toBeTruthy;
+        }
+        await deleteUser(userId);
+    });
+
+    it("should respond with 500 error due to pre-existing record", async () => {
+        const payload: UserRequestData = { userName: "guest123", password: "passwordpassword" };
+
+        const response = await request(app)
+                            .post("/api/users")
+                            .send(payload)
+                            .set("Accept", "application/json")
+                            .set("Content-Type", "application/json");
+
+        const user: UserRecord = response.body;
+        if (!user) {
+            throw new Error("invalid user");
+        }
+
+        const userId = user.id;
+        if (!userId) {
+            throw new Error("invalid userId");
+        }
+
+        const responseCopy = await request(app)
+                            .post("/api/users")
+                            .send(payload)
+                            .set("Accept", "application/json")
+                            .set("Content-Type", "application/json");
+
+        expect(responseCopy.status).toEqual(500);
         await deleteUser(userId);
     });
 });
